server/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2
django-appconf==1.1.0
django-cors-headers==4.7.0
django-filter==25.1
django-staticfiles==1.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.1.0
postgres==4.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
